Updated the functions for the numerical buttons.
Set displayText to a value of 0 and declared a variable for displayValue. Removed .innerHTML in the functions and set the value of the button as the displayText. Added a new line in each function that sets the displayValue variable as the displayText. Lastly I grabbed the element that has the class of displayText's innerHTML to the displayText variable to update the display.