Fixed issue if user clicks = with an empty current or previous operand.
Previously if the user clicked on the equals button while previous or current operand was empty, it would throw 'undefined' in the display. To combat that, I disabled the equals button for the user unless there is a value in both the previous and current operand using a if else statement to check for conditions.